fix article page tests
