+++ src/app/pages/ArticlePage/ArticlePage.jsx
@@ -1,0 +1,151 @@
+import React, { useContext } from 'react';
+import path from 'ramda/src/path';
+import styled from 'styled-components';
+import { string, node } from 'prop-types';
+import {
+  GEL_GROUP_1_SCREEN_WIDTH_MAX,
+  GEL_GROUP_2_SCREEN_WIDTH_MIN,
+  GEL_GROUP_3_SCREEN_WIDTH_MAX,
+  GEL_GROUP_4_SCREEN_WIDTH_MIN,
+  GEL_GROUP_4_SCREEN_WIDTH_MAX,
+  GEL_GROUP_5_SCREEN_WIDTH_MIN,
+} from '@bbc/gel-foundations/breakpoints';
+import {
+  GEL_MARGIN_ABOVE_400PX,
+  GEL_MARGIN_BELOW_400PX,
+  GEL_SPACING_DBL,
+  GEL_SPACING_TRPL,
+  GEL_SPACING_QUAD,
+  GEL_SPACING_QUIN,
+} from '@bbc/gel-foundations/spacings';
+import { articleDataPropTypes } from '#models/propTypes/article';
+import ArticleMetadata from '#containers/ArticleMetadata';
+import { ServiceContext } from '#contexts/ServiceContext';
+import headings from '#containers/Headings';
+import text from '#containers/Text';
+import image from '#containers/Image';
+import Blocks from '#containers/Blocks';
+import timestamp from '#containers/ArticleTimestamp';
+import { GridWrapper } from '#lib/styledGrid';
+import ATIAnalytics from '#containers/ATIAnalytics';
+import ChartbeatAnalytics from '#containers/ChartbeatAnalytics';
+import ComscoreAnalytics from '#containers/ComscoreAnalytics';
+import articleMediaPlayer from '#containers/ArticleMediaPlayer';
+import LinkedData from '#containers/LinkedData';
+import MostReadContainer from '#containers/MostRead';
+import MostReadSection from '#containers/MostRead/section';
+import MostReadSectionLabel from '#containers/MostRead/label';
+import {
+  getArticleId,
+  getHeadline,
+  getSummary,
+  getFirstPublished,
+  getLastPublished,
+  getAboutTags,
+  getArticleSection,
+  getMentions,
+  getLang,
+} from '#lib/utilities/parseAssetData';
+
+const componentsToRender = {
+  headline: headings,
+  subheadline: headings,
+  audio: articleMediaPlayer,
+  video: articleMediaPlayer,
+  text,
+  image,
+  timestamp,
+};
+
+const StyledMain = styled.main`
+  flex-grow: 1;
+`;
+
+const ArticlePageMostReadSection = styled(MostReadSection)`
+  @media (max-width: ${GEL_GROUP_1_SCREEN_WIDTH_MAX}) {
+    margin: 0 ${GEL_MARGIN_BELOW_400PX} ${GEL_SPACING_TRPL};
+  }
+  @media (min-width: ${GEL_GROUP_2_SCREEN_WIDTH_MIN}) and (max-width: ${GEL_GROUP_3_SCREEN_WIDTH_MAX}) {
+    margin: 0 ${GEL_MARGIN_ABOVE_400PX} ${GEL_SPACING_QUAD};
+  }
+  @media (min-width: ${GEL_GROUP_4_SCREEN_WIDTH_MIN}) and (max-width: ${GEL_GROUP_4_SCREEN_WIDTH_MAX}) {
+    margin: 0 ${GEL_MARGIN_ABOVE_400PX} ${GEL_SPACING_QUIN};
+  }
+  @media (min-width: ${GEL_GROUP_5_SCREEN_WIDTH_MIN}) {
+    width: 100%; /* Needed for IE11 */
+    margin: 0 auto ${GEL_SPACING_TRPL};
+    padding: 0 ${GEL_SPACING_DBL};
+    max-width: ${GEL_GROUP_5_SCREEN_WIDTH_MIN};
+  }
+`;
+
+const ArticlePage = ({ pageData, mostReadEndpointOverride }) => {
+  const { articleAuthor } = useContext(ServiceContext);
+  const headline = getHeadline(pageData);
+  const description = getSummary(pageData) || getHeadline(pageData);
+  const firstPublished = getFirstPublished(pageData);
+  const lastPublished = getLastPublished(pageData);
+  const aboutTags = getAboutTags(pageData);
+
+  const MostReadWrapper = ({ children }) => (
+    <ArticlePageMostReadSection>
+      <MostReadSectionLabel />
+      {children}
+    </ArticlePageMostReadSection>
+  );
+
+  MostReadWrapper.propTypes = {
+    children: node.isRequired,
+  };
+
+  return (
+    <>
+      <ATIAnalytics data={pageData} />
+      <ChartbeatAnalytics data={pageData} />
+      <ComscoreAnalytics />
+      <ArticleMetadata
+        articleId={getArticleId(pageData)}
+        title={headline}
+        author={articleAuthor}
+        firstPublished={firstPublished}
+        lastPublished={lastPublished}
+        section={getArticleSection(pageData)}
+        aboutTags={aboutTags}
+        mentionsTags={getMentions(pageData)}
+        lang={getLang(pageData)}
+        description={description}
+      />
+      <LinkedData
+        showAuthor
+        type="Article"
+        seoTitle={headline}
+        headline={headline}
+        datePublished={firstPublished}
+        dateModified={lastPublished}
+        aboutTags={aboutTags}
+      />
+      <StyledMain role="main">
+        <GridWrapper>
+          <Blocks
+            blocks={path(['content', 'model', 'blocks'], pageData)}
+            componentsToRender={componentsToRender}
+          />
+        </GridWrapper>
+      </StyledMain>
+      <MostReadContainer
+        mostReadEndpointOverride={mostReadEndpointOverride}
+        wrapper={MostReadWrapper}
+      />
+    </>
+  );
+};
+ArticlePage.propTypes = {
+  pageData: articleDataPropTypes.isRequired,
+  mostReadEndpointOverride: string,
+};
+
+ArticlePage.defaultProps = {
+  mostReadEndpointOverride: null,
+};
+
+export default ArticlePage;
